projeto/requirements.txt
gradio
openai
python-dotenv
requests
beautifulsoup4
python-dateutil
unidecode
fastapi
uvicorn